Move fs watch code to its own module
