+++ proxy/src/transport/tls/mod.rs
@@ -1,5 +1,4 @@
 // These crates are only used within the `tls` module.
-extern crate ring;
 extern crate rustls;
 extern crate tokio_rustls;
 extern crate untrusted;
